requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai>=1.3.7
bubbletea-chat[llm]
python-multipart>=0.0.6
pydantic>=2.6.0
python-dotenv>=1.0.0
aiohttp>=3.9.1